Add error handling and delete feature

+++ client-app/src/app/settings/feature/settings-my-advertisements/settings-my-advertisements.component.ts
@@ -1,15 +1,20 @@
 import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs';
+import { Observable, ReplaySubject } from 'rxjs';
+import { filter, switchMap, takeUntil, tap } from 'rxjs/operators';
 import { Advertisement } from 'src/app/advertisements/data-access/models/advertisement.model';
+import { showInfoBar } from 'src/app/data-access/info-bar/info-bar.actions';
+import { MessageService } from 'src/app/modules/core/services/message.service';
 import { ListViewType } from 'src/app/shared/ui/list-container/list-container.component';
+import { DestroyObservable } from 'src/app/shared/utils/destroy-observable';
 import {
     changeListViewType,
     deleteAdvertisement,
     loadAdvertisements,
     loadImages,
     selectAdvertisements,
+    selectError,
     selectListViewType,
 } from '../../data-access/store/settings-my-advertisements';
 
@@ -17,26 +22,55 @@ import {
     selector: 'app-settings-my-advertisements',
     templateUrl: './settings-my-advertisements.component.html',
     styleUrls: ['./settings-my-advertisements.component.scss'],
+    providers: [DestroyObservable],
 })
 export class SettingsMyAdvertisementsComponent implements OnInit {
     advertisements$: Observable<Advertisement[] | null> = this.store.select(selectAdvertisements);
-
     selectedListViewType$: Observable<ListViewType> = this.store.select(selectListViewType);
-    constructor(public dialog: MatDialog, private store: Store) {}
+    deleteAdvertisement$: ReplaySubject<number> = new ReplaySubject<number>(1);
+    error$: Observable<string | null> = this.store.select(selectError);
+    constructor(
+        private store: Store,
+        private router: Router,
+        private route: ActivatedRoute,
+        private messageService: MessageService,
+        private destroy$: DestroyObservable
+    ) {}
 
     ngOnInit(): void {
         this.store.dispatch(loadAdvertisements());
+        this.deleteAdvertisement$
+            .pipe(
+                takeUntil(this.destroy$),
+                switchMap((id) =>
+                    this.messageService.showConfirmationDialog('Czy na pewno chcesz usunąć to ogłoszenie?').pipe(
+                        filter(({ result }) => result),
+                        tap(() => {
+                            this.store.dispatch(deleteAdvertisement({ id }));
+                        })
+                    )
+                )
+            )
+            .subscribe();
+        this.error$
+            .pipe(
+                takeUntil(this.destroy$),
+                filter((error): error is string => error !== null),
+                tap((error) => this.store.dispatch(showInfoBar({ message: error, messageType: 'error' })))
+            )
+            .subscribe();
     }
 
     changeListViewType = (listViewType: ListViewType) => {
         this.store.dispatch(changeListViewType({ listViewType }));
     };
-
-    deleteAdvertisement = (id: number) => {
-        this.store.dispatch(deleteAdvertisement({ id }));
-    };
-
     loadImages = (advertisement: Advertisement) => {
         this.store.dispatch(loadImages({ itemId: advertisement.item.id }));
+    };
+    openAdvertisementForm = () => {
+        this.router.navigate(['add'], {
+            relativeTo: this.route,
+            queryParams: { returnUrl: this.router.routerState.snapshot.url },
+        });
     };
 }
